Add "Launch in a Jupyter notebook" button
Hi there, I’m Dan from Deepnote. We’re a small startup working on a new collaborative data science notebook, compatible with Jupyter. On thing we're really interested in is making it easier for people to try and learn from existing projects. I was trying out Deepnote with your notebooks for book chapters, and I thought it might be useful for others too to have a one-click button to run them, so I submitted a PR.
Hope you don't mind and find it useful!

Thanks for your work,
Dan